requirements_dev.txt
homeassistant
pytest-homeassistant-custom-component
pre-commit
